Refactoring the Repo based in new decisions
