--- src/classes/routine/routine.cpp
@@ -1,10 +1,0 @@
-/*
-* service.cpp
-* 
-* Author: Carlos Craveiro (@CarlosCraveiro)
-* Created On: September 11, 2021
-* 
-*/
-
-#include "routine.hpp"
-
