+++ src/classes/service/service.hpp
@@ -1,26 +1,33 @@
 /*
- * service.hpp
- *
- * Author: Carlos Craveiro (@CarlosCraveiro)
- * Created On: September 15, 2021
- */
+* routine.hpp
+* 
+* Author: Carlos Craveiro (@CarlosCraveiro)
+* Created On: September 11, 2021
+* 
+*/
 
 #pragma once
-#include <thread>
+#include <string>
 #include <mutex>
-#include "../routine-handler/routine-handler.hpp"
 
-class Service {
+class Status {
 	public:
-		 std::vector<RoutineHandler> routines;
-		 std::vector<EndpointHandler> endpoints;
-	public:
-		 nlohmann::json getServiceStatus(void);
-		nlohmann::json updateAll(nlohmann::json updateList);
-		nlohmann::json registerEndpoints(void);
-		nlohmann::json registerRoutines(void);
+		enum stateT { 
+			MISSINGDEPENDENCIES = 0, UNINITIALIZED, INITIALIZED, RUNNING, STOPED, DEAD
 
-		void buildEndpoint(/*IEndpoint& newEndpoint*/);
-		void buildRoutine(IRoutine& newRoutine);
+		};
+		stateT state;
+		std::mutex mtx;
+};
+
+class IRoutine {
+	public:
+		std::string id;
+	public:
+		virtual void onLoad()=0;
+		virtual void loop()=0;
+		virtual void unload()=0;
+		IRoutine(std::string id): id{id} {}
+		~IRoutine(void){};
 };
